Refactor components to use common directory structure and remove deprecated files
