+++ client/src/components/common/ErrorBoundary.tsx
