+++ client/src/components/common/ProtectedRoute.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Navigate } from 'react-router-dom'
 import { Box, CircularProgress, Typography } from '@mui/material'
-import { useAuth } from '../context/AuthContext'
+import { useAuth } from '../../context/AuthContext'
 
 interface ProtectedRouteProps {
   children: React.ReactNode
