+++ client/src/pages/Properties/CreateProperty.tsx
@@ -3,7 +3,7 @@ import { Box, Alert, CircularProgress } from '@mui/material'
 import { useNavigate, useParams } from 'react-router-dom'
 import { useTheme } from '@mui/material/styles'
 import { useCurrency, CURRENCY_CONFIG } from '../../hooks/useCurrency'
-import Sidebar from '../../components/basic/Sidebar'
+import Sidebar from '../../components/common/Sidebar'
 import Titlebar from '../../components/basic/Titlebar'
 import CustomButton from '../../components/basic/CustomButton'
 import ImageUploadSection from '../../components/properties/ImageUploadSection'
