+++ client/src/pages/Properties/Property.tsx
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from 'react'
 import { Box, Typography, Alert } from '@mui/material'
 import { useNavigate } from 'react-router-dom'
-import Sidebar from '../../components/basic/Sidebar'
+import Sidebar from '../../components/common/Sidebar'
 import Titlebar from '../../components/basic/Titlebar'
 import PropertyCard from '../../components/properties/PropertyCard'
 import PropertyFilters from '../../components/properties/PropertyFilters'
-import ErrorBoundary from '../../components/ErrorBoundary'
+import ErrorBoundary from '../../components/common/ErrorBoundary'
 import { Pagination, PropertyCardSkeletonGrid, SkipLink } from '../../components/common'
 import { getAllProperties, deleteProperty } from '../../services/property'
 import type { Property } from '../../services/property'
